chore(reusable component):
- Make table reusable by components that render data in a tabular form
- Make SignUp a functional component
