+++ app/client/test/components/searchResult.test.jsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import SearchResult from '../../components/search/SearchResult';
-import SearchResultTable from '../../components/tables/SearchResultTable';
+import { mount, shallow } from 'enzyme';
+import Table from '../../components/tables/Table';
+import { SearchResult } from '../../components/search/SearchResult';
 
+let props;
 let mountedSearchResult;
-const props = {
-  searchQuery: 'mazma'
-};
 const searchResult = () => {
   if (!mountedSearchResult) {
-    mountedSearchResult = shallow(
+    mountedSearchResult = mount(
       <SearchResult {...props} />
     );
   }
@@ -17,7 +15,38 @@ const searchResult = () => {
 };
 
 describe('<SearchResult />', () => {
-  it('should always render <SearchResultTable />', () => {
-    expect(searchResult().find(SearchResultTable).length).toBe(1);
+  beforeEach(() => {
+    props = {
+      searchUser: jest.fn(() => Promise.resolve()),
+      searchResult: {
+        users: [
+          {
+            id: 1,
+            firstName: 'Mary',
+            lastName: 'Mazi',
+            email: 'me@yahoo.com',
+            phoneNumber: '080987655342',
+            groups: []
+          }
+        ],
+        pagination: {
+          totalRows: 1
+        }
+      },
+      searchQuery: 'mazma'
+    };
+  });
+
+  it('should mount with handlePageClick()', () => {
+    const page = { selected: 1 };
+    const handlePageClickSpy = jest.spyOn(searchResult().instance(), 'handlePageClick');
+    searchResult().instance().handlePageClick(page);
+    expect(handlePageClickSpy).toHaveBeenCalledTimes(1);
+  });
+
+  it('should render search results if at least one user was found', () => {
+    const tableDisplay = searchResult().find(Table);
+    expect(tableDisplay.find('#firstName').text()).toBe('Mary');
+    expect(tableDisplay.find('#lastName').text()).toBe('Mazi');
   });
 });
